feat: remove unnecessary views
