--- app/src/main/java/com/techtown/tarsosdsp_pitchdetect/SongRecordActivity.java
@@ -1,16 +1,0 @@
-package com.techtown.tarsosdsp_pitchdetect;
-
-import androidx.appcompat.app.AppCompatActivity;
-import androidx.fragment.app.FragmentTransaction;
-
-import android.os.Bundle;
-
-public class SongRecordActivity extends AppCompatActivity {
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_song_record);
-
-    }
-}
