--- app/src/main/java/com/techtown/tarsosdsp_pitchdetect/WeakSentenceListRecordActivity.java
@@ -1,14 +1,0 @@
-package com.techtown.tarsosdsp_pitchdetect;
-
-import androidx.appcompat.app.AppCompatActivity;
-
-import android.os.Bundle;
-
-public class WeakSentenceListRecordActivity extends AppCompatActivity {
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_weak_sentence_list_record);
-    }
-}
