+++ app/src/main/java/com/techtown/tarsosdsp_pitchdetect/correction/PitchCorrection.java
@@ -14,7 +14,7 @@ import com.techtown.tarsosdsp_pitchdetect.R;
 import com.techtown.tarsosdsp_pitchdetect.global.CustomPitchCorrectionListDto;
 import com.google.android.material.bottomnavigation.BottomNavigationView;
 import com.google.android.material.navigation.NavigationBarView;
-import com.techtown.tarsosdsp_pitchdetect.MyRecordActivity;
+import com.techtown.tarsosdsp_pitchdetect.MyRecord.MyRecordActivity;
 import com.techtown.tarsosdsp_pitchdetect.SongListActivity;
 
 public class PitchCorrection extends AppCompatActivity {
