+++ app/src/main/java/com/techtown/tarsosdsp_pitchdetect/correction/RhythmCorrection.java
@@ -12,7 +12,7 @@ import android.widget.ListView;
 
 import com.google.android.material.bottomnavigation.BottomNavigationView;
 import com.google.android.material.navigation.NavigationBarView;
-import com.techtown.tarsosdsp_pitchdetect.MyRecordActivity;
+import com.techtown.tarsosdsp_pitchdetect.MyRecord.MyRecordActivity;
 import com.techtown.tarsosdsp_pitchdetect.R;
 import com.techtown.tarsosdsp_pitchdetect.SongListActivity;
 import com.techtown.tarsosdsp_pitchdetect.global.CustomPitchCorrectionListDto;
