+++ app/src/main/res/layout/activity_my_record.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/white"
-    tools:context=".MyRecordActivity">
+    tools:context=".MyRecord.MyRecordActivity">
 
     <TextView
         android:id="@+id/textView_myrecord"
