+++ app/src/main/res/layout/activity_weak_sentence_list.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/white"
-    tools:context=".WeakSentenceListActivity">
+    tools:context=".MyRecord.WeakSentenceListActivity">
 
 
     <TextView
@@ -22,12 +22,11 @@
         app:layout_constraintTop_toTopOf="parent" />
 
     <ListView
-        android:id="@+id/listView"
+        android:id="@+id/weak_listView"
         android:layout_width="329dp"
         android:layout_height="467dp"
-        android:layout_marginTop="220dp"
+        android:layout_marginTop="156dp"
         app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.5"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent">
 
@@ -41,7 +40,6 @@
         android:layout_marginBottom="14dp"
         android:gravity="center_vertical"
         android:orientation="vertical"
-        app:layout_constraintBottom_toTopOf="@+id/listView"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintHorizontal_bias="0.155"
         app:layout_constraintStart_toStartOf="parent"
@@ -49,7 +47,7 @@
         app:layout_constraintVertical_bias="0.8">
 
         <TextView
-            android:id="@+id/songTextView2"
+            android:id="@+id/weak_songTextView"
             android:layout_width="72dp"
             android:layout_height="33dp"
             android:text="title"
@@ -58,7 +56,7 @@
             android:textStyle="bold" />
 
         <TextView
-            android:id="@+id/singerTextView2"
+            android:id="@+id/weak_singerTextView"
             android:layout_width="70dp"
             android:layout_height="30dp"
             android:layout_marginLeft="2dp"
